feat(lecturer): create course
